Struct dynamic_reload::Lib

pub struct Lib {
    pub lib: Library,
    pub loaded_path: PathBuf,
    pub original_path: Option<PathBuf>,
}

Contains the information for a loaded library.

Fields

lib

The actual loaded library. Refer to the libloading documentation on how to use this.

loaded_path

This is the path from where the library was loaded (which may be in a temporary directory)

original_path

Original location of the file. This is keep so dynamic_reload knows which file to look for updates in case the library has been changed.