Enum dynamic_reload::UpdateState

pub enum UpdateState {
    Before,
    After,
    ReloadFalied(Error),
}

This is the states that the callback function supplied to update can be called with.

Variants

Before

Set when a shared library is about to be reloaded. Gives the application time to save state, do clean up, etc

After

Called when a library has been reloaded. Allows the application to restore state.

ReloadFalied

In case reloading of the library failed (broken file, etc) this will be set and allow the application to to deal with the issue.